requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
exceptiongroup==1.3.1
faiss-cpu==1.8.0
fastapi==0.112.0
ffmpy==1.0.0
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
gradio==4.42.0
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.37
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.26.4
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pillow==10.4.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
ruff==0.15.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.37.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.12.0
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.30.6
websockets==12.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
